requirements.txt
accelerate==0.23.0
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
clip==1.0
cloud-init==23.3.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
deluge==2.0.3
distro==1.4.0
distro-info==0.23+ubuntu1.1
entrypoints==0.3
filelock==3.12.4
fsspec==2023.9.2
ftfy==6.1.1
GeoIP==1.3.2
gpg==1.13.1
httplib2==0.14.0
huggingface-hub==0.17.3
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipython-genutils==0.2.0
Jinja2==2.10.1
joblib==1.3.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-core==4.6.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
more-itertools==4.2.0
mpmath==1.3.0
netifaces==0.10.4
networkx==3.1
nltk==3.7
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
peft==0.5.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.9.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.6
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-libtorrent==1.1.13
pytz==2023.3.post1
pyxdg==0.26
PyYAML==5.3.1
regex==2023.10.3
rencode==1.0.6
requests==2.22.0
requests-unixsocket==0.2.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sos==4.5.6
ssh-import-id==5.10
sympy==1.12
systemd-python==234
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchvision==0.11.2+cu111
tqdm==4.66.1
traitlets==4.3.3
transformers==4.34.0
triton==2.1.0
Twisted==18.9.0
typing-extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.8
xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
